+++ .gitignore
@@ -4,3 +4,4 @@ node_modules/
 .DS_Store
 .nyc_output
 yarn-error.log
+*.tgz
