+++ boot/package.json
@@ -1,6 +1,6 @@
 {
   "private": true,
-  "main": "../dist/cjs/boot.js",
+  "main": "../dist/es5/boot.js",
   "jsnext:main": "../dist/es2015/boot.js",
   "module": "../dist/es2015/boot.js",
   "types": "../dist/es2015/boot.d.ts"
