+++ examples/SSR/parcel-react-ssr/stream-server/middleware.js
@@ -4,7 +4,7 @@ import fs from 'fs';
 import MultiStream from 'multistream';
 import {getProjectStyles, createStyleStream} from 'used-styles';
 import {printDrainHydrateMarks, ImportedStream} from 'react-imported-component';
-import {createLoadableStream} from 'react-imported-component/server';
+import {createLoadableStream, createLoadableTransformer, getLoadableTrackerCallback} from 'react-imported-component/server';
 import React from 'react';
 import ReactDOM from 'react-dom/server';
 import {StaticRouter} from 'react-router-dom';
@@ -57,11 +57,23 @@ export default function middleware(req, res) {
   // create a style steam
   const styledStream = createStyleStream(projectStyles, (style) => {
     // emit a line to header Stream
-    headerStream.push(`<link href="dist/${style}" rel="stylesheet">\n`);
+    headerStream.push(`\n/*used style*/<link href="dist/${style}" rel="stylesheet">\n`);
   });
 
   // allow client to start loading js bundle
-  res.write(`<!DOCTYPE html><html><head><script defer src="${manifect['client.js']}"></script>`);
+  res.write(`<!DOCTYPE html><html><head><script defer src="${manifect['client.js']}"></script>\n`);
+
+  /*
+
+  res
+   <- headerStream
+   <- middleStream
+   <- styledStream
+      <- importedHelper
+         <- htmlStream (React)
+         <- htmlStream "end"
+   */
+
 
   const middleStream = readableString('</head><body><div id="app">');
   const endStream = readableString('</div></body></html>');
@@ -73,10 +85,15 @@ export default function middleware(req, res) {
     endStream,
   ];
 
+  const importedHelper = createLoadableTransformer(streamUID, getLoadableTrackerCallback("exampleTracker"));
+
   MultiStream(streams).pipe(res);
 
   // start by piping react and styled transform stream
-  htmlStream.pipe(styledStream, {end: false});
+  htmlStream
+    .pipe(importedHelper)
+    .pipe(styledStream, {end: false});
+
   htmlStream.on('end', () => {
     // push loaded chunks information
     headerStream.push(printDrainHydrateMarks(streamUID));
