+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-imported-component",
-  "version": "5.5.3",
+  "version": "6.0.0-1",
   "description": "I will import your component, and help to handle it",
   "main": "dist/es5/index.js",
   "jsnext:main": "dist/es2015/index.js",
@@ -79,6 +79,7 @@
   "files": [
     "dist",
     "boot",
-    "server"
+    "server",
+    "babel.js"
   ]
 }
